# agent_api_http/openapi-full.yaml
openapi: 3.1.0
info:
  title: UniCore HTTP API
  description: ''
  license:
    name: Apache 2.0
  version: 0.0.0-semantically-released
servers:
- url: http://localhost:3033
  description: Local development
paths:
...
  /auth/authorize:
    get:
      tags:
      - OAuth 2.0
      - Protocol
      summary: Authorize a pushed authorization request
      description: |-
        Starts the authorization of a pushed authorization request by reference, as defined by
        [RFC 9126](https://www.rfc-editor.org/rfc/rfc9126.html#name-authorization-request). The user agent is
        redirected to the consent page, or back to the client once consent has been given.
      operationId: auth_authorize
      parameters:
      - name: client_id
        in: query
        required: true
        schema:
          type: string
      - name: request_uri
        in: query
        required: true
        schema:
          type: string
      responses:
        '302':
          description: Redirect to the client's redirect URI
          headers:
            Location:
              schema:
                type: string
              description: The client's redirect URI
        '303':
          description: Redirect to the consent page
          headers:
            Location:
              schema:
                type: string
              description: The consent page URI
        '400':
          description: The query string is invalid
        '500':
          description: The authorization request could not be processed
  /auth/consent:
    get:
      tags:
      - OAuth 2.0
      - Protocol
      summary: Get the consent page
      description: Renders the HTML page on which the end-user grants or denies consent to a pushed authorization request.
      operationId: get_consent
      parameters:
      - name: request_uri
        in: query
        description: The `request_uri` returned by the pushed authorization request endpoint
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Consent page
          content:
            text/html:
              schema:
                type: string
        '400':
          description: The query string is invalid
        '500':
          description: The consent page could not be rendered
    post:
      tags:
      - OAuth 2.0
      - Protocol
      summary: Submit consent
      description: Records whether the end-user granted consent and redirects the user agent back to the client.
      operationId: post_consent
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ConsentForm'
        required: true
      responses:
        '302':
          description: Redirect to the client's redirect URI
          headers:
            Location:
              schema:
                type: string
              description: The client's redirect URI
        '415':
          description: The request body is not `application/x-www-form-urlencoded`
        '422':
          description: The request body is not a valid consent form
        '500':
          description: The consent could not be recorded
...
  /credential_offer:
    get:
      tags:
      - OpenID4VCI
      - Protocol
      summary: Receive a credential offer
      description: |-
        Receives a credential offer from a credential issuer, as defined by
        [OpenID4VCI](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#name-credential-offer-endpoint).
        `credential_offer` takes precedence when both parameters are present.
      operationId: credential_offer
      parameters:
      - name: credential_offer
        in: query
        description: JSON-encoded credential offer.
        required: false
        schema:
          type: string
      - name: credential_offer_uri
        in: query
        description: URL from which the credential offer can be retrieved.
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Credential offer received
        '400':
          description: Neither parameter is present, or the credential offer is invalid
        '500':
          description: The credential offer could not be stored
...
  /redirect:
    post:
      tags:
      - OID4VP / SIOPv2
      - Protocol
      summary: Submit an authorization response
      description: |-
        Receives and verifies the authorization response of a SIOPv2 or OID4VP authorization request, sent with the
        `direct_post` response mode.
      operationId: redirect
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AuthorizationResponseForm'
        required: true
      responses:
        '200':
          description: Authorization response verified
        '400':
          description: The authorization response is malformed or has no `state`
        '500':
          description: The authorization response could not be verified
...
components:
  schemas:
...
    AuthorizationResponseForm:
      oneOf:
      - $ref: '#/components/schemas/IdTokenAuthorizationResponse'
      - $ref: '#/components/schemas/VpTokenAuthorizationResponse'
      description: An authorization response sent to the relying party, either as a SIOPv2 ID Token or as an OID4VP VP Token.
...
    ConsentForm:
      type: object
      required:
      - client_id
      - request_uri
      - consent_given
      properties:
        client_id:
          type: string
        consent_given:
          type: boolean
        request_uri:
          type: string
...
    IdTokenAuthorizationResponse:
      type: object
      description: |-
        A [SIOPv2](https://openid.net/specs/openid-connect-self-issued-v2-1_0.html#name-self-issued-openid-provider-a)
        authorization response.
      required:
      - state
      - id_token
      properties:
        id_token:
          type: string
          description: The Self-Issued ID Token as a compact JWT.
        state:
          type: string
          description: The `state` of the authorization request, which identifies it.
...
    VpTokenAuthorizationResponse:
      type: object
      description: |-
        An [OID4VP](https://openid.net/specs/openid-4-verifiable-presentations-1_0.html#name-response) authorization
        response.
      required:
      - state
      - vp_token
      properties:
        state:
          type: string
          description: The `state` of the authorization request, which identifies it.
        vp_token:
          type: string
          description: JSON-encoded object that maps each DCQL credential query ID to a non-empty array of presentations.
  securitySchemes:
    access_token:
      type: http
      scheme: bearer
      description: Access token issued by the `/auth/token` endpoint.
...